Changement de groupe manuelle, routes guards frontend
- Implémentation du changement de groupe manuelle pour le gestionnaire de formation
- Mise en place de guardes de route directement au niveau du router et plus sur les pages dans un hook

EL OUAZIZI Walid

+++ front/src/stores/auth.js
@@ -53,8 +53,8 @@ export const useAuthStore = defineStore('authStore', () => {
             })
     }
 
-    const logout = async () => {
-        await api.delete(`/auth/logout`).catch(console.error)
+    const logout = async (makeRequest = true) => {
+        if(makeRequest) await api.delete(`/auth/logout`).catch(console.error)
         localStorage.removeItem('token')
         isLogged.value = false
         router.push('/login')
